ARUHA-1397 Move exceptions to a proper place
