+++ src/main/java/org/zalando/nakadi/controller/CursorsController.java
@@ -22,7 +22,7 @@ import org.zalando.nakadi.exceptions.NakadiException;
 import org.zalando.nakadi.exceptions.NoSuchEventTypeException;
 import org.zalando.nakadi.exceptions.ServiceUnavailableException;
 import org.zalando.nakadi.exceptions.UnableProcessException;
-import org.zalando.nakadi.exceptions.runtime.CursorUnavailableException;
+import org.zalando.nakadi.exceptions.runtime.CursorsAreEmptyException;
 import org.zalando.nakadi.exceptions.runtime.FeatureNotAvailableException;
 import org.zalando.nakadi.exceptions.runtime.RequestInProgressException;
 import org.zalando.nakadi.problem.ValidationProblem;
@@ -177,13 +177,6 @@ public class CursorsController {
         return Responses.create(Response.Status.SERVICE_UNAVAILABLE, ex.getMessage(), request);
     }
 
-    @ExceptionHandler({CursorUnavailableException.class, CursorsAreEmptyException.class})
-    public ResponseEntity<Problem> handleCursorsUnavailableException(final RuntimeException ex,
-                                                                     final NativeWebRequest request) {
-        LOG.debug(ex.getMessage(), ex);
-        return Responses.create(MoreStatus.UNPROCESSABLE_ENTITY, ex.getMessage(), request);
-    }
-
     @ExceptionHandler(RequestInProgressException.class)
     public ResponseEntity<Problem> handleRequestInProgressException(final RequestInProgressException ex,
                                                                     final NativeWebRequest request) {
@@ -203,12 +196,6 @@ public class CursorsController {
                                                            final NativeWebRequest request) {
         LOG.debug(ex.getMessage(), ex);
         return Responses.create(Problem.valueOf(Response.Status.NOT_IMPLEMENTED, "Feature is disabled"), request);
-    }
-
-    public static class CursorsAreEmptyException extends RuntimeException {
-        public CursorsAreEmptyException() {
-            super("Cursors are empty");
-        }
     }
 
 }
