+++ src/main/java/org/zalando/nakadi/exceptions/runtime/CursorsAreEmptyException.java
@@ -1,0 +1,7 @@
+package org.zalando.nakadi.exceptions.runtime;
+
+public class CursorsAreEmptyException extends MyNakadiRuntimeException1 {
+    public CursorsAreEmptyException() {
+        super("Cursors are empty");
+    }
+}
